+++ bin/uls_config/global_config.py
@@ -31,6 +31,7 @@ input_rerun_delay = 1                           # Time in seconds between rerun 
 input_disable_stderr = True                     # Enable STDERR output disabling (see value below to specify when this should happen)
 input_disable_stderr_after = 25                 # Disable stderr output after x input_cli cycles --> to prevent buffer overflow
 input_queue_size = 15000                        # Maximum number of events we want to store in-memory, default is 10000
+
     ## EAA
 bin_eaa_cli = "ext/cli-eaa/bin/akamai-eaa"
 eaa_cli_feeds = ['ACCESS', 'ADMIN', 'CONHEALTH', 'DEVINV', 'DIRHEALTH']
@@ -88,6 +89,9 @@ output_http_expected_status_code = 200          # Return Code for successful del
 output_http_liveness_check = True               # Send an OPTIONS request to probe the HTTP Server is live
 output_http_default_formattype = 'json-list'   # The default "formattype" being used in standard operation
 output_http_formattypes = ['json-list', 'single-event']     # List of choices (valid formattypes)
+output_http_compression = False                 # The default value to devide if HTTP compression is enabled or disabled by default
+output_http_default_compression_type = "GZIP"           # The default compression mechanism
+output_http_compression_choices = ['GZIP','BROTLI']      # A list of allowed compression methods
 
     ## FILE
 output_file_encoding = "utf-8"                  # FILE Encoding setting